amplicon: tweak target guessing + add more region data
Adjust thresholds and conditions when a guess is made or withheld and
tweak information in output files.

Added entries to the amplicon region table.  Some numbers are derived
from looking at results of running the amplicon target guessing and
looking up which primers were used and what the literature says what
regions those amplify (this goes for the mitochondrial 16S entry and the
euk 18S V1/V2 numbers, the rest of the 18S numbers are from the
glamr_data_management repo scripts)