Snakefile: fix input functions that depend on runinfo file
There are two input functions that depend on the content of the runinfo
file.  They need to know to the single/paired-endedness of the run.
Previously, the get_reads rule, which makes the runinfo files, was a
checkpoint.  This resulted in some truly strange errors in certain cases
where the runinfo file was returned as the input file.  Usually running
the pipeline from get_reads was fine, but starting from a later step
e.g. amplicon_guess_target triggered the error.

This commit creates an separate checkpoint (get_reads turns back into a
rule) that just link the runinfo file.  Input functions whose output
depend on runinfo's content should get the file from this checkpoint.