Snakefile: Raw reads download verification -- part 1 of 4
This commit is first in a series that adds verification of downloaded
raw read files.  The idea is that some checks are run against downloaded
files and that those checks can be done for already downloaded files, so
as not to trigger re-downloadding everything.

This commit moves the raw-read-stats file generation around and adds a
dummy check function (which currently never fails.)

The tricky part is that downstream rules should be independently ask for
the stats file or the raw fastqs.  And it must work in single-ended and
paired-ended cases.