Snakefile: properly pass the ncbi api key via environment
Note: If set, this must be a proper key for kingfisher to work, else, it
seems, kingfisher will go into an endless re-try loop with one minute
sleep.  Even setting NCBI_API_KEY to an empty string will fail with
kingfisher, so the variable must be unset if no real key is available.